+++ src/resource/solidDataset.ts
@@ -489,7 +489,7 @@ export async function deleteSolidDataset(
 }
 
 /**
- * Create an empty Container at the given URL.
+ * Create a Container at the given URL (which optionally can have it's only meta-data).
  *
  * Throws an error if creating the Container failed, e.g. because the current user does not have
  * permissions to, or because the Container already exists.
@@ -500,13 +500,15 @@ export async function deleteSolidDataset(
  *
  * @param url URL of the empty Container that is to be created.
  * @param options Optional parameter `options.fetch`: An alternative `fetch` function to make the HTTP request, compatible with the browser-native [fetch API](https://developer.mozilla.org/docs/Web/API/WindowOrWorkerGlobalScope/fetch#parameters).
+ * @param solidDataset Optional parameter - if provided we use this dataset as the body of the HTT request, meaning it's data is included in the Container resource.
  * @since 0.2.0
  */
 export async function createContainerAt(
   url: UrlString | Url,
   options: Partial<
     typeof internal_defaultFetchOptions
-  > = internal_defaultFetchOptions
+  > = internal_defaultFetchOptions,
+  solidDataset?: SolidDataset
 ): Promise<SolidDataset & WithServerResourceInfo> {
   url = internal_toIriString(url);
   url = url.endsWith("/") ? url : url + "/";
@@ -515,17 +517,55 @@ export async function createContainerAt(
     ...options,
   };
 
-  const response = await config.fetch(url, {
-    method: "PUT",
-    headers: {
-      Accept: "text/turtle",
-      "Content-Type": "text/turtle",
-      "If-None-Match": "*",
-      // This header should not be required to create a Container,
-      // but ESS currently expects it:
-      Link: `<${ldp.BasicContainer}>; rel="type"`,
-    },
-  });
+  let response;
+  let datasetToReturn;
+
+  if (solidDataset === undefined) {
+    response = await config.fetch(url, {
+      method: "PUT",
+      headers: {
+        Accept: "text/turtle",
+        "Content-Type": "text/turtle",
+        "If-None-Match": "*",
+        // This header should not be required to create a Container,
+        // but ESS currently expects it:
+        Link: `<${ldp.BasicContainer}>; rel="type"`,
+      },
+    });
+
+    datasetToReturn = createSolidDataset();
+  } else {
+    const datasetWithChangelog = internal_withChangeLog(solidDataset);
+
+    // If we are creating a new container, then it can't have existed before,
+    // so we can't be doing an 'update' to it. So if the provided dataset has
+    // updates then we assume there was a programmer error. (This may require
+    // changing based on user feedback (e.g., perhaps the dataset was read
+    // from a different container, we augmented that, and now want to use it
+    // as meta-data for our new container)).
+    if (isUpdate(datasetWithChangelog, url)) {
+      throw new Error(
+        `Attempting to create a new container with a dataset that contains updates!`
+      );
+    }
+
+    response = await config.fetch(url, {
+      method: "PUT",
+      body: await triplesToTurtle(
+        toRdfJsQuads(solidDataset).map(getNamedNodesForLocalNodes)
+      ),
+      headers: {
+        Accept: "text/turtle",
+        "Content-Type": "text/turtle",
+        "If-None-Match": "*",
+        // This header should not be required to create a Container,
+        // but ESS currently expects it:
+        Link: `<${ldp.BasicContainer}>; rel="type"`,
+      },
+    });
+
+    datasetToReturn = solidDataset;
+  }
 
   if (internal_isUnsuccessfulResponse(response)) {
     if (
@@ -537,8 +577,9 @@ export async function createContainerAt(
       return createContainerWithNssWorkaroundAt(url, options);
     }
 
+    const containerType = solidDataset === undefined ? "empty" : "non-empty";
     throw new FetchError(
-      `Creating the empty Container at [${url}] failed: [${response.status}] [${response.statusText}].`,
+      `Creating the ${containerType} Container at [${url}] failed: [${response.status}] [${response.statusText}].`,
       response
     );
   }
@@ -547,7 +588,7 @@ export async function createContainerAt(
   const containerDataset: SolidDataset &
     WithChangeLog &
     WithServerResourceInfo = freeze({
-    ...createSolidDataset(),
+    ...datasetToReturn,
     internal_changeLog: { additions: [], deletions: [] },
     internal_resourceInfo: resourceInfo,
   });